fixing qt no longer accepting floats in place of ints bug

+++ app/editor/event_editor/event_properties.py
@@ -853,4 +853,4 @@ class CommandDelegate(QStyledItemDelegate):
             painter.drawText(left, top + font_height, command)
             font.setPointSize(prev_size)
             painter.setFont(font)
-            painter.drawLine(left, top + 1.25 * font_height, right, top + 1.25 * font_height)
+            painter.drawLine(left, int(top + 1.25 * font_height), right, int(top + 1.25 * font_height))
